--- Tests/CommandRegistryTests/XCTestManifests.swift
@@ -1,2 +1,0 @@
-import XCTest
-
